video-handler/requirements.txt
aio_pika
pyyaml